+++ CHANGELOG.md
@@ -26,6 +26,11 @@
 * Add uploadNdkUnityLibraryMappings flag to bugsnag extension
   [#306](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/306)
 
+## 5.2.0 (2020-09-22)
+
+* Support mapping upload from Android Gradle Plugin through a proxy
+  [#298](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/298)
+
 ## 5.1.0 (2020-09-17)
 
 * Remove uploadDebugBuildMappings flag in favour of variantFilter